+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# v0.3.1
+
+- Fixed bug that caused server to return 500 when given empty string for phone number
+    - Given phone number now must be a valid American phone number
+    - Formatting and non digit characters are ignored
+
 # v0.3.0
 
 - /mail route now uses CORS policy.